modularizing errors
