+++ src/errors/changelog.errors.js
@@ -1,120 +1,6 @@
 import { MAX_ENTRY_LENGTH } from "../config/constants.js";
 
-/**
- * Represents an error during the extraction of data from a GitHub Pull Request.
- */
-export class PullRequestDataExtractionError extends Error {
-  /**
-   * Constructs the PullRequestDataExtractionError instance.
-   * @param {string} [message="Error extracting data from Pull Request"] - Custom error message.
-   */
-  constructor(message = "Error extracting data from Pull Request") {
-    super(message);
-    this.name = this.constructor.name;
-    /**
-     * Indicates whether this error should trigger a comment in the pull request.
-     * @type {boolean}
-     */
-    this.shouldResultInPRComment = false;
-  }
-}
 
-/**
- * Represents an error during the retrieval of content from a GitHub repository.
- */
-export class GetGithubContentError extends Error {
-  /**
-   * Constructs the GetGithubContentError instance.
-   * @param {string} [message="Error retrieving content from GitHub"] - Custom error message.
-   */
-  constructor(message = "Error retrieving content from GitHub repository") {
-    super(message);
-    this.name = this.constructor.name;
-    /**
-     * Indicates whether this error should trigger a comment in the pull request.
-     * @type {boolean}
-     */
-    this.shouldResultInPRComment = false;
-  }
-}
-
-/**
- * Represents an error during the creation of a file in a GitHub repository.
- */
-export class CreateFileError extends Error {
-  /**
-   * Constructs the CreateChangesetFileError instance.
-   * @param {string} [message="Error creating file in repository"] - Custom error message.
-   */
-  constructor(message = "Error creating file in repository") {
-    super(message);
-    this.name = this.constructor.name;
-    /**
-     * Indicates whether this error should trigger a comment in the pull request.
-     * @type {boolean}
-     */
-    this.shouldResultInPRComment = false;
-  }
-}
-
-/**
- * Represents an error during the update of a file in a GitHub repository.
- */
-export class UpdateFileError extends Error {
-  /**
-   * Constructs the UpdateFileError instance.
-   * @param {string} [message="Error updating file in repository"] - Custom error message.
-   */
-  constructor(message = "Error updating file in repository") {
-    super(message);
-    this.name = this.constructor.name;
-    /**
-     * Indicates whether this error should trigger a comment in the pull request.
-     * @type {boolean}
-     */
-    this.shouldResultInPRComment = false;
-  }
-}
-
-/**
- * Represents an error during the deletion of a file in a GitHub repository.
- */
-export class DeleteFileError extends Error {
-  /**
-   * Constructs the DeleteFileError instance.
-   * @param {string} [message="Error deleting file in repository"] - Custom error message.
-   */
-  constructor(message = "Error deleting file in repository") {
-    super(message);
-    this.name = this.constructor.name;
-    /**
-     * Indicates whether this error should trigger a comment in the pull request.
-     * @type {boolean}
-     */
-    this.shouldResultInPRComment = false;
-  }
-}
-
-/**
- * Represents an error that occurs when updating the label of a pull request.
- */
-export class UpdatePRLabelError extends Error {
-  /**
-   * Constructs the UpdatePRLabelError instance.
-   * @param {string} [message="There was an error updating the label of the pull request. Please ensure the PR is accessible and the label format is correct."] - Custom error message.
-   */
-  constructor(
-    message = "There was an error updating the label of the pull request. Please ensure the PR is accessible and the label format is correct."
-  ) {
-    super(message);
-    this.name = this.constructor.name;
-    /**
-     * Indicates whether this error should trigger a comment in the pull request.
-     * @type {boolean}
-     */
-    this.shouldResultInPRComment = false;
-  }
-}
 
 /**
  * Represents an error for a missing or malformed changelog heading in a PR description.
