+++ src/tests/errors/index.js
@@ -1,0 +1,3 @@
+export * from "./changelog.errors.test.js";
+export * from "./content.errors.test.js";
+export * from "./pullRequest.errors.test.js";
